requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
click==8.1.8
colorama==0.4.6
email-validator==2.1.1
fastapi==0.115.12
h11==0.14.0
idna==3.10
iso8601==2.1.0
pydantic==2.11.1
pydantic-settings==2.8.1
pydantic_core==2.33.0
pypika-tortoise==0.5.0
python-dotenv==1.1.0
pytz==2025.2
sniffio==1.3.1
starlette==0.46.1
tortoise-orm==0.24.2
typing-inspection==0.4.0
typing_extensions==4.13.0
uvicorn==0.29.0
